Add several components and models for Pokemon functionality.
The added components include InputWithButton, TypeFilter, DropdownMenu, and a ThemeProvider. Moreover, the models added include a Button model and a Dictionary model. Integration with an API is established for Pokemon data, and error handling for error and loading states is set up. Additionally, translations for English and French language are implemented in respective dictionaries.